More work on ChunkScreen updates
