+++ common/src/main/resources/assets/ftbchunks/lang/en_us.json
@@ -122,6 +122,7 @@
 	"ftbchunks.gui.admin_mode_disabled": "Admin Mode: Disabled",
 	"ftbchunks.gui.admin_mode_enabled": "Admin Mode: Enabled",
 	"ftbchunks.gui.admin_mode_info": "When enabled you can modify chunks, regardless of ownership",
+	"ftbchunks.gui.unclaim_all": "Unclaim All",
 	"ftbteamsconfig.ftbchunks": "FTB Chunks Properties",
 	"ftbteamsconfig.ftbchunks.allow_fake_players": "Allow All Fake Players",
 	"ftbteamsconfig.ftbchunks.allow_fake_players.tooltip": "Treat ALL fake players as allies of the team.\nWARNING: Setting this to true could allow hostile players to access your claims via any fake player. Set this to false if you're unsure.",
